+++ packages/i18n-core/CHANGELOG.md
@@ -1,3 +1,10 @@
+# v2.0.0
+
+- Added interpolation feature
+  - String wrapped with `[]` can be interpolated
+  - Added escape HTML feature to prevent XSS. `true` by default
+- Create move query params logic to `appendQueryParams` function in utils
+
 # v1.0.0
 
 - Created i18n class as an entry point for translation